Move error and loading to message directory
