+++ pages/_app.js
@@ -4,9 +4,9 @@ import { PointElement, Tooltip, registry } from "chart.js"
 import dynamic from "next/dynamic"
 
 import { useCensusEmbedded, CensusEmbeddedProvider } from "@components/Contexts/CensusEmbeddedContext"
-import Error_ from "@components/Error_"
-import Loading from "@components/Loading"
 import MainLayout from "@components/MainLayout"
+import Error_ from "@components/Message/Error_"
+import Loading from "@components/Message/Loading"
 import Sidebar from "@components/Navigation/Sidebar/Sidebar"
 import { Setup } from "@components/Setup"
 import { useBasicFetch, useFetchRuns } from "@utils/fetch"
